rgw/metadata: encapsulate role metadata handler in rgw_role.cc
