+++ src/rgw/driver/posix/rgw_sal_posix.cc
@@ -20,6 +20,7 @@
 #include <unistd.h>
 #include "rgw_multi.h"
 #include "include/scope_guard.h"
+#include "common/errno.h"
 
 #define dout_subsys ceph_subsys_rgw
 #define dout_context g_ceph_context
