+++ src/rgw/driver/rados/roles.h
@@ -25,6 +25,7 @@ namespace ceph { class Formatter; }
 class DoutPrefixProvider;
 class optional_yield;
 struct rgw_raw_obj;
+struct RGWRoleInfo;
 
 
 namespace rgwrados::roles {
@@ -34,7 +35,7 @@ int add(const DoutPrefixProvider* dpp,
         optional_yield y,
         librados::Rados& rados,
         const rgw_raw_obj& obj,
-        const rgw::sal::RGWRoleInfo& role,
+        const RGWRoleInfo& role,
         bool exclusive, uint32_t limit);
 
 /// Look up a role's id by name in the list.
